refactor(eslint-rules): inline base-hook-detector into each rule, drop utils/
Premature shared-utils abstraction: only two rules existed and the
duplicated surface area is ~18 lines (BASE_HOOK_NAME_PATTERN, the
BaseHookFunction type, getFunctionInit). Inlining makes each rule
self-contained at the cost of trivial duplication.

- Move all base-hook-detector exports into base-hook-signature.ts
  (signature rule owns the full pair-detection mechanism).
- Duplicate BASE_HOOK_NAME_PATTERN / BaseHookFunction / getFunctionInit
  into base-hook-no-forbidden-runtime.ts.
- Delete tools/eslint-rules/utils/ entirely.
